include datasource

+++ src/main/java/com/tales/miniautorizador/dto/TransacaoRequest.java
@@ -6,11 +6,11 @@ import java.math.BigDecimal;
 
 @Data
 public class TransacaoRequest {
-    private Integer numCartao;
+    private Long numCartao;
     private String senha;
     private BigDecimal valor;
 
-    public TransacaoRequest(Integer numCartao, String senha, BigDecimal valor) {
+    public TransacaoRequest(Long numCartao, String senha, BigDecimal valor) {
         this.numCartao = numCartao;
         this.senha = senha;
         this.valor = valor;
